backend/python/requirements.txt
flask
pyyaml
requests
pydantic
flask_cors